make sample_rate f32 and remove new method from vst

+++ reverb/src/allpass_filter.rs
@@ -6,7 +6,7 @@ pub struct AllpassFilter {
 }
 
 impl AllpassFilter {
-  pub fn new(sample_rate: f64) -> Self {
+  pub fn new(sample_rate: f32) -> Self {
     Self {
       delay_line: DelayLine::new((sample_rate * 0.015) as usize, sample_rate),
     }
